refactor: convert to a hybrid ESM/CommonJS package
