+++ package.json
@@ -10,44 +10,54 @@
   "bugs": {
     "url": "https://github.com/zwave-js/zwave-js-server/issues"
   },
-  "main": "dist/lib/index.js",
-  "bin": {
-    "zwave-server": "dist/bin/server.js",
-    "zwave-client": "dist/bin/client.js"
-  },
   "publishConfig": {
     "access": "public"
+  },
+  "type": "module",
+  "bin": {
+    "zwave-server": "dist-esm/bin/server.js",
+    "zwave-client": "dist-esm/bin/client.js"
+  },
+  "exports": {
+    ".": {
+      "import": "./dist-esm/lib/index.js",
+      "require": "./dist-cjs/lib/index.js"
+    },
+    "./package.json": "./package.json"
   },
   "scripts": {
     "lint": "eslint",
     "lint:fix": "eslint --fix && prettier -w .",
-    "test": "prettier --check src && tsc --noEmit && npm run lint && ts-node src/test/integration.ts",
+    "test": "prettier --check src && tsc --noEmit && npm run lint && tsx src/test/integration.ts",
     "build": "tsc -p .",
+    "postbuild": "esm2cjs --in dist-esm --out dist-cjs -l error -t node18",
     "prepare": "npm run build",
-    "prepublishOnly": "rm -rf dist && npm run build"
+    "prepublishOnly": "rm -rf dist && npm run build",
+    "try": "testbin"
   },
   "author": "",
   "license": "Apache-2.0",
   "dependencies": {
     "@homebridge/ciao": "^1.1.7",
     "minimist": "^1.2.8",
-    "ws": "^8.13.0"
+    "ws": "^8.18.0"
   },
   "peerDependencies": {
     "zwave-js": "^14.0.0"
   },
   "devDependencies": {
+    "@alcalzone/esm2cjs": "^1.4.0",
     "@eslint/eslintrc": "^3.1.0",
     "@eslint/js": "^9.14.0",
     "@tsconfig/node18": "^18.2.1",
     "@types/minimist": "^1.2.2",
     "@types/node": "^22.5.0",
     "@types/triple-beam": "^1.3.2",
-    "@types/ws": "^8.5.5",
+    "@types/ws": "^8.5.13",
     "@types/yargs": "^17.0.24",
     "@typescript-eslint/eslint-plugin": "^8.14.0",
     "@typescript-eslint/parser": "^8.14.0",
-    "alcalzone-shared": "^4.0.8",
+    "alcalzone-shared": "^5.0.0",
     "eslint": "^9.14.0",
     "eslint-config-prettier": "^9.1.0",
     "globals": "^15.9.0",
@@ -55,7 +65,7 @@
     "lint-staged": "^15.0.2",
     "prettier": "^3.0.0",
     "semver": "^7.5.4",
-    "ts-node": "^10.9.2",
+    "tsx": "^4.19.2",
     "typescript": "^5.3.3",
     "zwave-js": "^14.0.0"
   },
