+++ src/lib/const.ts
@@ -1,3 +1,5 @@
+import { createRequire } from "node:module";
+const require = createRequire(import.meta.url);
 export const version = require("../../package.json").version;
 
 // minimal schema version the server supports
