+++ src/lib/index.ts
@@ -1,4 +1,4 @@
-export * from "./command";
-export * from "./state";
-export { ZwavejsServer } from "./server";
-export const serverVersion: string = require("../../package.json").version;
+export * from "./command.js";
+export * from "./state.js";
+export { ZwavejsServer } from "./server.js";
+export { version as serverVersion } from "./const.js";
