+++ src/lib/message_handler.ts
@@ -1,5 +1,5 @@
-import { IncomingMessage } from "./incoming_message";
-import { ResultTypes } from "./outgoing_message";
+import { IncomingMessage } from "./incoming_message.js";
+import { ResultTypes } from "./outgoing_message.js";
 
 export interface MessageHandler {
   handle(message: IncomingMessage): Promise<ResultTypes[keyof ResultTypes]>;
