+++ src/lib/multicast_group/outgoing_message.ts
@@ -1,6 +1,6 @@
 import { VirtualValueID } from "zwave-js";
-import { MulticastGroupCommand } from "./command";
-import { SetValueResultType } from "../common";
+import { MulticastGroupCommand } from "./command.js";
+import { SetValueResultType } from "../common.js";
 
 export interface MulticastGroupResultTypes {
   [MulticastGroupCommand.setValue]: SetValueResultType;
