+++ src/lib/state.ts
@@ -46,7 +46,7 @@ import {
   ValueMetadataNumeric,
   ValueMetadataString,
 } from "@zwave-js/core";
-import { numberFromLogLevel } from "../util/logger";
+import { numberFromLogLevel } from "../util/logger.js";
 
 type Modify<T, R> = Omit<T, keyof R> & R;
