+++ src/lib/utils/incoming_message.ts
@@ -1,6 +1,6 @@
 import { RSSI } from "zwave-js";
-import { IncomingCommandBase } from "../incoming_message_base";
-import { UtilsCommand } from "./command";
+import { IncomingCommandBase } from "../incoming_message_base.js";
+import { UtilsCommand } from "./command.js";
 
 export interface IncomingCommandUtilsBase extends IncomingCommandBase {}
